requirements.txt
Cmake
web3==7.6.0
eth-account==0.13.4
fastapi==0.115.8
langchain==0.3.17
langchain_community==0.3.16
langchain_core==0.3.33
langchain_openai<=0.3.0
langgraph==0.2.69
orjson==3.10.15
pandas==2.2.3
pydantic==2.10.6
python-dotenv==1.0.1
Requests==2.32.3
uvicorn==0.34.0
faiss-cpu==1.10.0

pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
